feat: 신용평가 모델 연동

+++ src/main/java/com/billit/credit/dto/IncomeProofData.java
@@ -10,5 +10,5 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 @AllArgsConstructor
 public class IncomeProofData {
-    private BigDecimal income;
+    private float income;
 }
